Added Loadout Copy feature. So now when a player is in the loadout configuration screen and they are on a loadout they can hit copy and have a new loadout generated based on the current loadout.
Modified Dialog_ManageLoadouts to support the new button (and the 2 lines of code that make it work).

Modified MapComp_LoadoutManager to add a method overload to generate a unique label from a given string.

Modified Loadout and LoadoutSlot to support a Copy() method and a static Copy(object) method to handle making copies.