Added ability for pawns to fulfill loadout needs from animals. Also from prisoners...
Modified JobGiver_UpdateLoadout to look inside inventories of other pawns on the same map as the pawn with the loadout missing items once the other two item searches fail.

Added a new JobDriver TakeFromOther which handles the task of getting the pawn to the pawn with the item they want, reserving it, a brief pause and then taking the item.

If the item being taken is flagged to be equipped then uses CE's code to swap the item from inventory to equipped.