Did a little touch up and testing... Tried to limit how many pawn inventories are checked. Also tried to use Pawn.inventory.GetInnerContainer() method instead of Pawn.inventory.innerContainer field following a suggestion that the field was probably intended to be private.
My knowledge of how things get optimized in .net is still very basic so not sure if what I did to limit the inventories checked helps or if it ends up neutral.

Basically after both searches of the map come up empty for an item it will now locate all pawns on the same map which are either pack animals of the same faction as the searching pawn or prisoners hosted by the same faction as the searching pawn... as long as their inventories are not empty.  It will then try to fetch an item which matches the desired def much like if it found a hit before.