Added Generics to LoadoutSlots...
LoadoutSlots come in two flavors, those with a thingDef property (that's not null).  These are Specific in that they point to a particular item.  These are pretty much good old LoadoutSlots from before.

The new one has a property of genericDef which stores a few details connected to generics and to make things a little easier in the code by hiding some of what a generic is when it doesn't matter.  The key property of a Generic is it's lambda which is a Predicate<ThingDef> that defines the types of items that it covers (will try to pickup/drop).

Another new property of LoadoutSlots is the LoadoutCountType (enum) which has 2 states currently, PickupDrop and DropExcess.  At the moment this property is always available for a LoadoutSlot however it only really comes into play currently for Generics.  PickupDrop means the slot will try to fill a request as well as drop extras while DropExtra means ignore any amount of a matching item up to Count and then start dropping.

The behavior between Specifics and Generics are:
-When picking up: Specifics are deferred to first while maintaining Loadout Slot order.  If the sum of all items that fit specifics covered by the generic are less than what the generic wants, more are picked up.  If all the specifics add up to more than the generic, nothing happens.
-When dropping: Handling is currently incomplete as it still only considers one thing at a time and Generics need to consider multiple things...

The coverage idea behind mixed specifics/generics is that if you want a specific Item that may not be available the generic can fill the space.  Then later when the specific item becomes available the pawn fetches the specific and if the total count falls under the generic it may drop something the generic had picked up.

The other big concept of generics is to cover things where a pawn might automatically fetch something such as a meal (pawns like to keep one on hand), some raw food (to train an animal), or some drugs (to fill a drug policy).  By default those are DropExcess.

I've still got a few more things to sort out for Generics.  This includes a few bits for HoldTracker as well, currently in a separate file to keep it easier on me.