Added HoldTracker concept as well as finishing most of Generics. At this point the code is mostly done.  There are a few things that still need/would like to have done but it's functional and what is there is ready for further testing.
LoadoutSlot-Specific is a ThingDef and count.  Just like it used to be and there should be very little changed in functionality here.

LoadoutSlot-Generic is a genericDef with count and countType.  The genericDef basically has a lambda expression indicating what ThingDefs should get caught up by the slot.  A single slot can cover multiple items.  The count functions just like LoadoutSlot-Specific except there can be more than one ThingDef being counted.  countType indicates if the slot should try to pickup items or if items should only be dropped when in excess.

Unlike my first attempt all slots attempt to be fully filled so if there are only FMJ pistol rounds and you have a Specific slot looking for 10 and a Generic Pistol Ammo slot the pawn will try to collect 20 rounds.

The HoldTracker is a new concept which is surprisingly similar to the old LoadoutSlot system.  Basically if a pawn is instructed to pick something up and they have a non-empty loadout defined then the HoldTracker remembers ThingDef and count.

The HoldTracker functions vaguely like a LoadoutSlot-Generic set to DropExcess in that it won't trigger any pickups but will drop excess after all LoadoutSlots have been considered.

Additionally updated Harmony (It's not expected this will change in interface significantly) and modified the existing Harmony code to try and act as a guideline for how new patches should be done.  There are still some questions I haven't figured out but for MOST cases I've got a decent understanding of Harmony now.