Added some optimizations to ManageLoadouts UI.
Specifically found an optimization I forgot about for Generic Ammo so the predicate can fail out sooner.

Also spread generic testing (if there is an item on the map satisfying the generic) out over time so that in any given frame only one generic is being tested.

Also modified Generic and Specific testing to instead of finding EVERY item on the map which satisfies it it will stop on the FIRST item that is satisfied so performance should be more consistent on very busy maps rather than slowing down.