+Generate Loadout from Inventory & MoteThrower to JobDriver_Reload
New button (suggestion of button from Karim) on Pawn's gear tab which creates a new Loadout from that Pawn's inventory.  Button only shown if the Pawn has no/empty Loadout and has equipment/inventory items.

Possible refinement: Only show the button if the items in inventory are not marked by Generics that are DropExcess only.

Also moved the motethrower associated with reloading into the JobDriver and out of two other sections of code.  May throw an exception if the creator of the job did insufficient error checking.