Further separated out functions of CompAmmoUser.TryStartReload()
The Unload mechanic has been separated off (CompAmmoUser.Unload()) so that it can be called on it's own though TryStartReload() still takes a bool indicating to only perform an unload.  More work necessary to finish separating that off.  Though TryStartReload does error checking that Unload() does not.

Also Unload() will drop any items on the ground that didn't fit into inventory instead of destroying them.

Added CompAmmoUser.GetReloadJob() which takes a bool indicating if the job should be flagged PlayerForced or not.  This doesn't do any error checking and expects the caller to check for errors (atm).

This also sets the storedTarget and storedJobDef to null so the caller doesn't have to do that, if they should have some object stored in them the caller can do that after getting the Job.

More to do...