Some more tweaks on CompAmmoUser.TryStartReload.
Changed some logic around to clean things up a little.

Renamed Unload() to TryUnload().  Returns false if there is something wrong with the weapon (would indicate can't do the reload job).  Returns true if the weapon is already unloaded or if the unload was successful.

Added an warning message if the ammo was destroyed because it couldn't be dropped.

Changed GiveJob() to TryGiveJob() and modified it's logic as well.

Modified TryStartReload() so that it no longer takes a bool.  Instead those wishing to unload a weapon would call TryUnload().  Modified all in project found references to CompAmmoUser.TryStartReload(true) to instead use CompAmmoUser.TryUnload().

Testing suggests behavior has been maintained.  Turrets and Pawns, if clip is empty, initiate a reload before aiming/warmup and resume the tasks they were doing before.  Pawns will initiate a reload when not otherwise busy when their clip is low on ammo.

Untested: Ammo in a gun not equipped gets reloaded.