Fixed issue #120
Or at least hopefully fixed #120 regarding the red error text.  Not sure about the other issues as I haven't been able to replicate.  Pawns with multiple weapons and ammo definitions seem to pick things up fine.

The issue seemed to be caused by a failure to check if the weapons in the inventory rangedWeaponCache had CompAmmoUser before trying to use said comp.