Attempt to fix various issues related to drafted pawns auto-firing.
Issue 1: The code that inserted the call in IL was inserting it in the wrong place, breaking the key logic of the function.

Issue 2: The call to Verb_LaunchProjectileCE.GetPartialCoverBetween() had the 2nd and 3rd arguments backwards from the correct order, so was checking if the shooter had cover and if so the target was invalid (heh).

Issue 3: An older bug in GetPartialCoverBetween where the checkedCells list wasn't being updated with the cells which had been checked potentially reducing performance in all situations related to shooting.