Partial fix for pawns not auto shooting targets and optimization.
Added an optimization to Verb_LaunchProjectileCE to get rid of the cell list and instead use an object with very fast access time to check if a cell has been checked before (rather than walking a list each check).

Partially fixed a bug related to pawns not shooting at targets if the target has ANY cover instead of too much cover.

Disabled the old Harmony-AttackTargetFinder_BestAttackTarget_Patch as not only was it bugged (denying targets with ANY cover) but was also the wrong approach.

Added a new Harmony-JobDriverWait-CheckForAutoAttack_Patch which replaces the null argument to BestShootTargetFromCurrentPosition() with a predicate which will validate targets via Verb_LaunchProjectileCE.CanHitTargetFrom().