Reverted Verb_LaunchProjectileCE, Cleaned up new Harmony patch.
Used git to view an older version of Verb_LaunchProjectileCE.GetPartialCoverBetween() and undid the change I made to add a static call to that method.

Also removed the old harmony patch.  While interesting it's no longer useful.

Cleaned up Harmony-JobDriverWait-CheckForAutoAttack_Patch.cs as I was all over the map trying to figure out how to store a new local variable when it turns out all I needed was to replace an ldnull with an ldloc_s <verb> and a call to a method to generate the predicate.