Fixed a save/load issue (red text) on turrets, also job failure.
There was some red error text on save load related to turrets, fixed.

Took a fair bit of back and forth with NIA but also fixed the issue with pawns dropping out of their job when reloading a turret (basically while going to the ammo to pick it up or when returning to the turret with the ammo in hand).