Implemented an infix to change TendUtility to orderby BleedRate... ThenBy Severity.
Basically replaced part of/all of an original LinQ statement with a new one.  So less surgical than I'd like but born of my lack of knowledge and that the simpler approach was available.

Took some study since this was the first time I really looked at LinQ in IL.  If JUST BleedRate then would have been a simple prefix patch (helps to look at the code in a crummy debugger instead of the nicely patched ILSpy).