Harmony check... (#185)
* Converted Detours_VerbTracker to Harmony patch.

In pure IL (for the heck of it and because it was a short patch, also for speed since the alternative would involve an extra call).

Also fixed up my remarks from the previous harmony work.

* Converted Detours_HediffComp_GetsOld to Harmony.

Also some cleanup of my previous harmony patches (an unused variable still being declared...)

Also a new utility function to safely cast nullable bool to bool, returns false if the nullable bool has no value (null) and the value of the bool if it has a value.

* Converted Detours_Pawn_HealthTracker to Harmony.

Also a new utility method in HarmonyBase.

* Converted Detours_TradeDeal to Harmony.

* Converted the last of the Detours...
Pawn_EquipmentTracker (a series of postfixes and one infix that is basically a method call conversion)
MassUtility (a postfix patch as that was fastest to do and fit fine).

Some more utility stuff in HarmonyBase, including a large remarked block.

* Examined WorkGiver_InteractAnimal and Verb_TryStartCastOn patches.

Harmony-WorkGiver_InteractAnimal didn't really need any work, the target notation (HarmonyPatch) was a tad off but the code was fine.

Harmony-Verb_TryStartCastOn wasn't fundamentally changed, modified it to use more recent patch procedures but it didn't really need any work.

* Examined GenRadial.cs, nothing was needed to be done...

Except to modify it to use the new mechanic for running custom patches.  The infix code and such was basically unmodified.

Refactored somewhat to fit the pattern of patches.

* Inspected the remaining patches, no obvious work needed.

Should be all patched up to date.  There may be some things that need a little more work but this should be functional.

* Tried to fix my goof on the merge...
Also turned off harmony debug.