Attempted fix to pawns getting stuck unloading their inventory.
The issue was that the jobgiver was unaware of loadouts and so would re-issue jobs that were closed successfully.

It was a bit weird in that they didn't always get stuck, suppose a race condition of sorts where sometimes the unloadinventory flag would get cleaned and others it wouldn't.

The patch proved a tad harder than expected since I had thought I needed an OR condition but I actually needed an AND.