Mostly worked on holdtracker ammo system disabled issues...
CompAmmoUser - Added a check if the thing trying to be reloaded is a turret and if so pass the reload job on... this code path is untested though as it turns out it could never run...

JobDriver_Reload* - altered the inspect strings of the jobs so as to avoid NRE when ammo system disabled.  JobDriver_Reload is re-patched since somehow git decided the fix should be removed in a different commit.

JobGiver_CheckReload has been modified to handle auto-reloads if the ammo system is inactive.  Also modified to perform auto reloads if the pawn does NOT have a loadout but does have a weapon and the right ammo in inventory.  The changes made to support ammoSystem being off made it easy to extend to no loadouts.

WorkGiver_Reload Turret - removed a line of code that was remarked in JobOnThing so that this will again trigger reloads of turrets when ammo system is off.

Utility_HoldTracker methods relating to building a picture of the pawn's inventory (ie to figure out what to drop) now consider if the ammo system is disabled and will not try to count ammo in the guns a pawn may have equipped or in inventory.