Converted LoadoutManager class to GameComponent.
This involved making most of the class instance based rather than static.  Also had to learn a few things about how GameComponent works.

This commit has HoldTracker data not saved/loaded since I'm stuck on an odd problem with that.  Is save compatible in that you can load your save game but any Loadouts (and HoldTracker data) in the save will be dropped.