Modified HoldRecords to more closely follow Loadouts.
This fixed the save/load issue I was having.  There might have been another solution but I knew this path would work (since Loadouts worked) so I didn't pursue that other path.

Also this commit returns LoadoutManager to a mostly static interface.

Since there is a window where LoadoutManager should not be returning/accepting actual data there are checks in place to see if it's apt to return something and if not, doesn't return anything.

This is because there is a risk of data bleeding from one instance to the next when loading a save game while an active game is running.