Fixed turrets not reloading if something was blocking them
As per issue #174 turrets can't be reloaded if their cell has an obstruction in it. As a workaround, pawns no longer place ammo down at all and reload from CarryTracker. This might need to be changed in the future if we ever attempt to fix another issue where turrets can't be reloaded properly if their magazine size is larger than the ammo's stack count.

Although that's not currently an issue, if someone wanted to say, add a Minigun turret with 1200 rounds of magazine capacity they'd have a problem.