Modified PawnColumnWorkers to be a bit more unified.
PawnColumnWorker_Loadout contains data that is key for all 3 (Outfits, Drugs, and Loadouts) to function in the assign tab.  That information also makes it relatively easy to tweak the look of all 3 PawnColumnWorkers.

A new translation string was added as I couldn't find a fitting RimWorld string to use.  The transpiler has items that should be alterable without messing the code up near the top. (apparelString and drugString)

Renamed the QuickAndDirty cs file to something a bit more logical and cleaned up the original code a bit.  Added the new button replacement code to this file in a new class.

PawnColumnWorker_Loadout.DoCell() contains a variety of remarks which are helpful in maintaining the transpiler for the other buttons.  Additionally due to complexity the new transpiler includes an overview of the patch operations and that overview is referenced in the code. So the PawnColumnWorker_Loadout.DoCell() combined with the overview and references informs how the transpiler works.

The transpiler could have been made *FAR* simpler by allowing the old code to run and then inserting calls to alter the local variable as well as method replacers... the transpiler should result in much faster IL code than that would have done and this is a frequently called UI element (basically every frame that Assign tab is open), so went with a bit more complexity to gain speed.

A small amount of additional complexity comes in because the changes to the IL were repeated 3 times across two classes in the same way so the patch code was built to be able to loop on itself.